% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB_ks_map.R
\name{ks_map}
\alias{ks_map}
\title{Find kinase-substrate relationships}
\usage{
ks_map(
  protData_filename,
  psiteData_filename,
  uniprot_gn_filename,
  fudge_factor,
  ksNetwork_filename,
  sub_norm = T,
  working_dir,
  ks_outputName
)
}
\arguments{
\item{protData_filename}{file name of protein data frame}

\item{psiteData_filename}{file name of phosphorylation data frame}

\item{uniprot_gn_filename}{file name of the uniprot accession number and gene name matching info, downloadable from our website}

\item{fudge_factor}{a factor between 0 and 1, default to 0.01 to be added to the entire data for robustness}

\item{ksNetwork_filename}{file name of known kinase-substrate relationships, downloadable from our website}

\item{working_dir}{the directory output files will be deposited in}

\item{ks_outputName}{a label for the mapped relationships}
}
\description{
Find kinase-substrate relationships
}
\examples{
ks_map
}
\keyword{map}
