% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB_prep_prot_psite.R
\name{protData_prep}
\alias{protData_prep}
\title{convert to log2 scale}
\usage{
protData_prep(protData_filename, fudge_factor = 0.01)
}
\arguments{
\item{protData_filename}{file name of protein data frame}

\item{fudge_factor}{a factor between 0 and 1, default to 0.01 to be added to the entire data for robustness}
}
\description{
convert to log2 scale
}
\examples{
protData_prep
}
\keyword{data}
\keyword{prepare}
