% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB_prep_prot_psite.R
\name{psiteData_prep}
\alias{psiteData_prep}
\title{convert to log2 scale and normalise psites with mother protein abundance}
\usage{
psiteData_prep(
  psiteData_filename,
  uniprot_gn_filename,
  fudge_factor,
  protData_p
)
}
\arguments{
\item{psiteData_filename}{file name of phosphorylation data frame}

\item{uniprot_gn_filename}{file name of the uniprot accession number and gene name matching info, downloadable from our website}

\item{fudge_factor}{a factor between 0 and 1, default to 0.01 to be added to the entire data for robustness}

\item{protData_p}{processed protein data, the output object of protData_prep function}
}
\description{
convert to log2 scale and normalise psites with mother protein abundance
}
\examples{
protData_prep
}
\keyword{data}
\keyword{prepare}
